starting out
goal is to create an extension that replaces certain keywords with asterisks, etc.

Found a block of code that can help me replace stuff, but need couldn't understand why the code was formatted that way, so did googling. apparently it has to do with jquery. (the dollar sign function,etc) more research will be done to gain an understanding before continuing